terrain: Draw single-pass terrain+decal geometry indexed
Convert terrainDecalVBO from 12 duplicated vertices per tile
(non-indexed, 4 triangles) to 5 shared vertices per tile (4 corners +
center) plus a static u32 index buffer forming the same center-fanned
triangles. Vertex memory drops ~58% and vertex shader invocations per
tile from 12 to 5.

Decal tangents were previously computed per-triangle on the duplicated
vertices. With shared vertices, each vertex accumulates its
tangent/bitangent over the tile triangles it belongs to before
orthogonalization (identical on uniform-slope tiles, smooth instead of
faceted on curved ones). Degenerate decal UVs are skipped instead of
feeding NaN tangents into the accumulation.

The combined pass now draws via the same addDrawRangeElements batching
as the depth pass, so the sector-size clamp accounts for its
5-vertices-per-tile requirement, TerrainCombinedTemplate's (previously
unused) index type widens to u32 for whole-map index values, and an
empty map (no decal buffers) skips the pass.

+++ lib/ivis_opengl/gfx_api.h
@@ -1096,7 +1096,7 @@ namespace gfx_api
 	};
 
 	template<REND_MODE render_mode, SHADER_MODE shader>
-	using TerrainCombinedTemplate = typename gfx_api::pipeline_state_helper<rasterizer_state<render_mode, DEPTH_CMP_LEQ_WRT_OFF, 255, polygon_offset::disabled, stencil_mode::stencil_disabled, cull_mode::back>, primitive_type::triangles, index_type::u16,
+	using TerrainCombinedTemplate = typename gfx_api::pipeline_state_helper<rasterizer_state<render_mode, DEPTH_CMP_LEQ_WRT_OFF, 255, polygon_offset::disabled, stencil_mode::stencil_disabled, cull_mode::back>, primitive_type::triangles, index_type::u32,
 	std::tuple<TerrainCombinedUniforms>,
 	std::tuple<
 	vertex_buffer_description<sizeof(TerrainDecalVertex), gfx_api::vertex_attribute_input_rate::vertex, // TerrainDecalVertex struct
